refactor and add redis
