+++ handlers/__init__.py
@@ -1,3 +1,4 @@
 from . import start
 from . import parse_group
-from . import statistics_and_analysis
+from . import statistics
+from . import tops
